Moved EstimationTask and EstimationFeature as a separate component. Same components would be used for both expanded/contracted view which would be driven by expanded attribute
